+++ src/modules/centerdevice/mod.rs
@@ -64,4 +64,4 @@ impl<'a> TryFrom<&'a CenterDeviceConfig> for AuthorizedClient<'a> {
     }
 }
 
-sub_module!("centerdevice", "Access CenterDevice from the CLI", auth, delete, download, search, upload);
+sub_module!("centerdevice", "Access CenterDevice from the CLI", auth, delete, download, search, upload, users);